requirements.txt
Flask==2.3.3
groq==0.25.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
email-validator==2.0.0.post2
python-dotenv==1.0.0
Werkzeug==2.3.7
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
python-jose==3.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
openai==0.28.1
python-docx==0.8.11
PyPDF2==3.0.1
SpeechRecognition==3.10.0
pyttsx3==2.90
python-dateutil==2.8.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.6.1
python-docx==0.8.11
python-pptx==0.6.21
itsdangerous>=2.1.2
Flask-Mail==0.10.0
ratelimit==2.2.1
